requirements.txt
requests==2.31.0
Pillow==9.4.0
pandas==2.1.1
statistics
plotly

